Add minimal specification
This is the smallest possible delta compared to String#replace that I could think of. This approach takes the _searchValue_ and turns it into a global regular expression if it isn't already, carefully escaping special characters when needed. It throws when RegExp.prototype[Symbol.replace] is not callable (e.g. when it has been deleted).

This way, we avoid the introduction of RegExp.prototype[Symbol.replaceAll] which would basically do the same thing as RegExp.prototype[Symbol.replace] for global regular expressions anyway.